+++ spec/grape/endpoint_spec.rb
@@ -254,6 +254,37 @@ describe Grape::Endpoint do
     end
   end
 
+  describe '#declared; call from child namespace' do
+    before do
+      subject.format :json
+      subject.namespace :something do
+        params do
+          requires :id, type: Integer
+        end
+        resource ':id' do
+          params do
+            requires :foo
+            optional :bar
+          end
+          get do
+            {
+              params: params,
+              declared_params: declared(params)
+            }
+          end
+        end
+      end
+    end
+
+    it 'should include params defined in the parent namespace' do
+      get '/something/123', foo: 'test', extra: 'hello'
+      expect(last_response.status).to eq 200
+      json = JSON.parse(last_response.body, symbolize_names: true)
+      expect(json[:params][:id]).to eq 123
+      expect(json[:declared_params].keys).to include :foo, :bar, :id
+    end
+  end
+
   describe '#params' do
     it 'is available to the caller' do
       subject.get('/hey') do
@@ -464,7 +495,7 @@ describe Grape::Endpoint do
       end
 
       get '/hey'
-      last_response.status.should == 403
+      last_response.status.should == 500
       last_response.body.should == "This is not valid."
     end
 
@@ -486,6 +517,16 @@ describe Grape::Endpoint do
       get '/hey.json'
       last_response.status.should == 403
       last_response.body.should == '{"dude":"rad"}'
+    end
+
+    it 'can specifiy headers' do
+      subject.get '/hey' do
+        error!({ 'dude' => 'rad' }, 403, 'X-Custom' => 'value')
+      end
+
+      get '/hey.json'
+      last_response.status.should == 403
+      last_response.headers['X-Custom'].should == 'value'
     end
   end